backend/requirements.txt
bcrypt==4.0.1
docx2txt==0.9
fastapi==0.136.0
groq==1.1.2
httpx==0.28.1
pandas==3.0.2
passlib==1.7.4
pdfplumber==0.11.9
pgvector==0.4.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
SQLAlchemy==2.0.49
uvicorn==0.44.0
